ws: Move the remotectl command into the ws directory
This is how its distributed, and how it logically fits.

The prompting factor is translations and long term packaging.
