+++ po/POTFILES.in
@@ -3,10 +3,10 @@
 # List of source files outside of pkg/ containing translatable strings.
 src/bridge/cockpitdbussetup.c
 src/bridge/cockpitpackages.c
-src/common/cockpitcertificate.c
 src/common/cockpitconnect.c
 src/ws/cockpit.appdata.xml.in
 src/ws/cockpit.desktop.in
+src/ws/cockpitcertificate.c
 src/ws/cockpithandlers.c
 pkg/lib/credentials.js
 pkg/lib/patterns.js
