+++ po/POTFILES.in.in
@@ -1,8 +1,8 @@
 [encoding: UTF-8]
 # List of source files outside of pkg/ containing translatable strings.
 src/bridge/cockpitdbussetup.c
-src/common/cockpitcertificate.c
 src/common/cockpitconnect.c
 src/ws/cockpit.appdata.xml.in
 src/ws/cockpit.desktop.in
+src/ws/cockpitcertificate.c
 src/ws/cockpithandlers.c
