+++ src/bridge/cockpitconnect.h
