+++ src/common/Makefile-common.am
@@ -33,8 +33,6 @@ EXTRA_DIST += \
 noinst_LIBRARIES += libcockpit-common.a
 
 libcockpit_common_a_SOURCES = \
-	src/common/cockpitcertificate.c \
-	src/common/cockpitcertificate.h \
 	src/common/cockpitconnect.c \
 	src/common/cockpitconnect.h \
 	src/common/cockpitenums.h \
