--- src/remotectl/Makefile-remotectl.am
@@ -1,20 +1,0 @@
-if WITH_COCKPIT_WS
-
-sbin_PROGRAMS += remotectl
-
-remotectl_SOURCES = \
-	src/remotectl/certificate.c \
-	src/remotectl/remotectl.c \
-	src/remotectl/remotectl.h \
-	$(NULL)
-
-remotectl_CFLAGS = \
-	$(COCKPIT_CFLAGS) \
-	$(NULL)
-
-remotectl_LDADD = \
-	libcockpit-common.a \
-	$(COCKPIT_LIBS) \
-	$(NULL)
-
-endif
